templates/starter-project-adk/requirements.txt
google-adk>=0.2
pyyaml>=6.0
pytest>=8.0
httpx>=0.27
